% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetFormula.r
\name{glmnet}
\alias{glmnet}
\alias{glmnet.default}
\alias{glmnet.formula}
\alias{predict.glmnet.formula}
\alias{coef.glmnet.formula}
\alias{print.glmnet.formula}
\title{Formula interface for elastic net modelling with glmnet}
\usage{
glmnet(x, ...)

\method{glmnet}{default}(x, y, ...)

\method{glmnet}{formula}(formula, data, alpha = 1, ..., weights = NULL,
  offset = NULL, subset = NULL, na.action = getOption("na.action"),
  drop.unused.levels = FALSE, xlev = NULL, sparse = FALSE,
  use.model.frame = FALSE)

\method{predict}{glmnet.formula}(object, newdata, offset = NULL,
  na.action = na.pass, ...)

\method{coef}{glmnet.formula}(object, ...)

\method{print}{glmnet.formula}(x, digits = max(3, getOption("digits") - 3),
  print.deviance.ratios = FALSE, ...)
}
\arguments{
\item{x}{For the default method, a matrix of predictor variables.}

\item{...}{For \code{glmnet.formula} and \code{glmnet.default}, other arguments to be passed to \link[glmnet:glmnet]{glmnet::glmnet}; for the \code{predict} and \code{coef} methods, arguments to be passed to their counterparts in package glmnet.}

\item{y}{For the default method, a response vector or matrix (for a multinomial response).}

\item{formula}{A model formula; interaction terms are allowed and will be expanded per the usual rules for linear models.}

\item{data}{A data frame or matrix containing the variables in the formula.}

\item{alpha}{The elastic net mixing parameter. See \link[glmnet:glmnet]{glmnet::glmnet} for more details.}

\item{weights}{An optional vector of case weights to be used in the fitting process. If missing, defaults to an unweighted fit.}

\item{offset}{An optional vector of offsets, an \emph{a priori} known component to be included in the linear predictor.}

\item{subset}{An optional vector specifying the subset of observations to be used to fit the model.}

\item{na.action}{A function which indicates what should happen when the data contains missing values. For the \code{predict} method, \code{na.action = na.pass} will predict missing values with \code{NA}; \code{na.omit} or \code{na.exclude} will drop them.}

\item{drop.unused.levels}{Should factors have unused levels dropped? Defaults to \code{FALSE}.}

\item{xlev}{A named list of character vectors giving the full set of levels to be assumed for each factor.}

\item{sparse}{Should the model matrix be in sparse format? This can save memory when dealing with many factor variables, each with many levels.}

\item{use.model.frame}{Should the base \link[=model.frame]{model.frame} function be used when constructing the model matrix? This is the standard method that most R modelling functions use, but has some disadvantages. The default is to avoid \code{model.frame} and construct the model matrix term-by-term; see \link[=glmnet.model.matrix]{discussion}.}

\item{object}{For the \code{predict} and \code{coef} methods, an object of class \code{glmnet.formula}.}

\item{newdata}{For the \code{predict} method, a data frame containing the observations for which to calculate predictions.}

\item{digits}{Significant digits in printed output.}

\item{print.deviance.ratios}{Whether to print the table of deviance ratios, as per \link[glmnet:print.glmnet]{glmnet::print.glmnet}.}
}
\description{
Formula interface for elastic net modelling with glmnet
}
\details{
The \code{glmnet} function in this package is an S3 generic with a formula and a default method. The former calls the latter, and the latter is simply a direct call to the \code{glmnet} function in package glmnet. All the arguments to \code{glmnet::glmnet} are (or should be) supported.

The code works in a similar manner to \code{lm}, \code{glm} and other modelling functions. The arguments are used to generate a \emph{model frame}, which is a data frame augmented with information about the roles the columns play in fitting the model. This is then turned into a \emph{model matrix} and a response vector, which are passed to \code{glmnet::glmnet} along with any arguments in \code{...}. If \code{sparse} is TRUE, then \code{Matrix::sparse.model.matrix} is used instead of \code{stats::model.matrix} to create the model matrix.

The \code{predict} and \code{coef} methods are wrappers for the corresponding methods in the glmnet package. The former constructs a predictor model matrix from its \code{newdata} argument and passes that as the \code{newx} argument to \code{glmnet:::predict.glmnet}.
}
\section{Value}{

For \code{glmnet.formula}, an object of class \code{glmnet.formula}. This is basically the same object created by \code{glmnet::glmnet}, but with extra components to allow formula usage.
}

\examples{
glmnet(mpg ~ ., data=mtcars)

glmnet(Species ~ ., data=iris, family="multinomial")

\dontrun{

# Leukemia example dataset from Trevor Hastie's website
download.file("http://web.stanford.edu/~hastie/glmnet/glmnetData/Leukemia.RData",
              "Leukemia.RData")
load("Leukemia.Rdata")
leuk <- do.call(data.frame, Leukemia)
glmnet(y ~ ., leuk, family="binomial")
}
}
\seealso{
\link[glmnet:glmnet]{glmnet::glmnet}, \link[glmnet:predict.glmnet]{glmnet::predict.glmnet}, \link[glmnet:coef.glmnet]{glmnet::coef.glmnet}, \link[=model.frame]{model.frame}, \link[=model.matrix]{model.matrix}
}
