% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glmnetUtils.r
\docType{package}
\name{glmnetUtils}
\alias{glmnetUtils}
\alias{glmnetUtils-package}
\title{A package to provide the familiar formula/data frame interface for \code{glmnet}.}
\description{
A package to provide the familiar formula/data frame interface for \code{glmnet}.
}

