% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvAlpha.r
\name{plot.cvAlpha.glmnet}
\alias{plot.cvAlpha.glmnet}
\alias{plotLambda}
\title{Plot the elastic net cross-validation results by alpha and lambda}
\usage{
\method{plot}{cvAlpha.glmnet}(x, ...)

plotLambda(x, ...)
}
\arguments{
\item{x}{An object of class \code{cvAlpha.glmnet}.}

\item{...}{Other arguments to be passed to \code{plot}.}
}
\description{
Plot the elastic net cross-validation results by alpha and lambda
}
\details{
The plot method for \code{cvAlpha.glmnet} objects plots the average cross-validated loss by lambda, for each value of alpha. Each line represents one \code{cv.glmnet} fit, corresponding to one value of alpha. Note that the specific lambda values can vary substantially by alpha, hence no attempt is made to put them on a common scale. Instead, the lines are simply the cross-validated loss results in sequential order, with the smallest lambda values for each fit on the left.

The \code{plotLambda} function gives the best (lowest) cross-validated loss for each value of alpha.
}
\seealso{
\code{\link{cvAlpha.glmnet}}, \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet}}, \code{\link{plot}}
}

