% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvAlpha.r
\name{predict.cvAlpha.glmnet}
\alias{coef.cvAlpha.glmnet}
\alias{predict.cvAlpha.glmnet}
\alias{predict.cvAlpha.glmnet.formula}
\title{Get predictions from a crossvalidated cvAlpha.glmnet object}
\usage{
\method{predict}{cvAlpha.glmnet}(object, newx, alpha, which = match(TRUE,
  abs(object$alpha - alpha) < 1e-08), ...)

\method{predict}{cvAlpha.glmnet.formula}(object, newdata, alpha,
  which = match(TRUE, abs(object$alpha - alpha) < 1e-08),
  na.action = na.pass, ...)

\method{coef}{cvAlpha.glmnet}(object, alpha, which = match(TRUE,
  abs(object$alpha - alpha) < 1e-08), ...)
}
\arguments{
\item{object}{An object of class \code{cvAlpha.glmnet}, or which inherits from it.}

\item{newx}{A matrix of predictor variables.}

\item{alpha}{The value of alpha at which to compute predictions. This must match with the values of alpha supplied to \code{cvAlpha.glmnet}.}

\item{which}{An alternative way of specifying alpha; the index number of the desired value within the alpha vector. If both \code{which} and \code{alpha} are supplied, the former takes precedence.}

\item{...}{Further arguments to be passed to \code{glmnet:::predict.cv.glmnet}.}

\item{newdata}{A data frame containing the observations for which to calculate predictions.}
}
\description{
Get predictions from a crossvalidated cvAlpha.glmnet object
}
\details{
The \code{predict} method computes predictions for a specific alpha value given a \code{cvAlpha.glmnet} object. It looks up the supplied alpha (possibly supplied indirectly via the \code{which} argument) in the object's stored \code{alpha} vector, and calls \code{glmnet:::predict.cv.glmnet} on the corresponding \code{cv.glmnet} fit. All the arguments to that function are (or should be) supported.

The \code{coef} method is similar, returning the coefficients for the selected alpha value via \code{glmnet:::coef.cv.glmnet}.
}
\examples{
predict(cvAlpha.glmnet(mpg ~ ., data=mtcars), mtcars, alpha=1)

\dontrun{

# Leukemia example dataset from Trevor Hastie's website
download.file("http://web.stanford.edu/~hastie/glmnet/glmnetData/Leukemia.RData",
              "Leukemia.RData")
load("Leukemia.Rdata")
leuk <- do.call(data.frame, Leukemia)
leuk.cva <- cvAlpha.glmnet(y ~ ., leuk, family="binomial")
pred <- predict(leuk.cva, leuk, which=6)
}
}
\seealso{
\code{\link[glmnet:predict.cv.glmnet]{glmnet:::predict.cv.glmnet}}, \code{\link[glmnet:coef.cv.glmnet]{glmnet:::coef.cv.glmnet}}
}

